requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy==1.23.5
scipy==1.15.2
scikit-image==0.21.0
tqdm==4.67.0
kornia==0.7.3
kornia-augmentation==0.7.0
timm==1.0.15
PyYAML==6.0.2
tensorboard==2.14.0
matplotlib==3.7.1
Pillow==10.0.0
tifffile==2023.9.18
